[chore] 뷰 분리
