+++ TeamIntroduce/TeamIntroduce/Sources/Presnetaion/MemberDetail/MemberDetailView.swift
@@ -41,123 +41,6 @@ struct MemberDetailView: View {
     }
 }
 
-struct MemberProfileView: View {
-    let profile: MemberProfile
-
-    var body: some View {
-        HStack {
-            Spacer()
-
-            VStack(spacing: 7) {
-                Image(systemName: profile.image)
-                    .font(.system(size: 40))
-
-                Text(profile.name)
-                    .pretendardFont(family: .Regular, size: 13)
-
-                Text(profile.role)
-                    .pretendardFont(family: .Regular, size: 13)
-                    .foregroundStyle(.textSecondary100)
-
-
-                Text(profile.mbti)
-                    .pretendardFont(family: .Regular, size: 13)
-                    .padding(.horizontal, 12)
-                    .padding(.vertical, 6)
-                    .background(Capsule().fill(.gray20))
-            }
-
-            Spacer()
-        }
-        .cardStyle()
-    }
-}
-
-struct IntroductionView: View {
-    let profile: MemberProfile
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            Text("자기소개")
-                .pretendardFont(family: .SemiBold, size: 14)
-            Text(profile.introduction)
-                .pretendardFont(family: .Regular, size: 13)
-                .foregroundStyle(.textSecondary100)
-                .lineSpacing(5)
-                .cardStyle()
-        }
-    }
-}
-
-struct StrengthsView: View {
-    let profile: MemberProfile
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            Text("주요장점")
-                .pretendardFont(family: .SemiBold, size: 14)
-            VStack {
-                ForEach(profile.strengths, id: \.self) { item in
-                    HStack(alignment: .firstTextBaseline, spacing: 8) {
-                        Circle().frame(width: 5, height: 5)
-                        Text(item)
-                            .pretendardFont(family: .Regular, size: 13)
-                            .foregroundStyle(.textSecondary100)
-                            .fixedSize(horizontal: false, vertical: true)
-                    }
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                }
-            }
-            .cardStyle()
-        }
-    }
-}
-
-struct CollabStyleView: View {
-    let profile: MemberProfile
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            Text("협업스타일")
-                .pretendardFont(family: .SemiBold, size: 14)
-            Text(profile.collabStyle)
-                .pretendardFont(family: .Regular, size: 13)
-                .foregroundStyle(.textSecondary100)
-                .lineSpacing(5)
-                .cardStyle()
-        }
-    }
-}
-
-struct BlogView: View {
-    let profile: MemberProfile
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            Text("블로그")
-                .pretendardFont(family: .SemiBold, size: 14)
-
-            Text(profile.blogURL)
-                .pretendardFont(family: .Regular, size: 13)
-                .cardStyle()
-
-        }
-    }
-}
-
-struct EditButton: View {
-    var body: some View {
-        Button(action: {}) {
-            Text("수정")
-                .foregroundColor(.white)
-        }
-        .frame(maxWidth: .infinity)
-        .frame(height: 48)
-        .background(.backGroundPrimary)
-        .clipShape(RoundedRectangle(cornerRadius: 16))
-    }
-}
-
 struct MemberDetailView_Previews: PreviewProvider {
     static var previews: some View {
         MemberDetailView()
