+++ TeamIntroduce/TeamIntroduce/Sources/Presnetaion/MemberDetail/MemberProfileView.swift
@@ -1,0 +1,39 @@
+//
+//  MemberProfileView.swift
+//  TeamIntroduce
+//
+//  Created by 김민희 on 8/12/25.
+//
+import SwiftUI
+
+struct MemberProfileView: View {
+    let profile: MemberProfile
+
+    var body: some View {
+        HStack {
+            Spacer()
+
+            VStack(spacing: 7) {
+                Image(systemName: profile.image)
+                    .font(.system(size: 40))
+
+                Text(profile.name)
+                    .pretendardFont(family: .Regular, size: 13)
+
+                Text(profile.role)
+                    .pretendardFont(family: .Regular, size: 13)
+                    .foregroundStyle(.textSecondary100)
+
+
+                Text(profile.mbti)
+                    .pretendardFont(family: .Regular, size: 13)
+                    .padding(.horizontal, 12)
+                    .padding(.vertical, 6)
+                    .background(Capsule().fill(.gray20))
+            }
+
+            Spacer()
+        }
+        .cardStyle()
+    }
+}
